feat(terminal): submit on route, improved code
